+++ MANIFEST.in
@@ -59,6 +59,7 @@ include psutil/arch/freebsd/cpu.c
 include psutil/arch/freebsd/disk.c
 include psutil/arch/freebsd/init.h
 include psutil/arch/freebsd/mem.c
+include psutil/arch/freebsd/pids.c
 include psutil/arch/freebsd/proc.c
 include psutil/arch/freebsd/proc_socks.c
 include psutil/arch/freebsd/sensors.c
